All Runtimes: Now sizeof(tick)=4 on all platforms (unsigned int).

+++ targets/Linux/plc_Linux_main.c
@@ -25,7 +25,7 @@
 #define _LogError(text,...) _Log(LOG_CRITICAL, text, ##__VA_ARGS__)
 #define _LogWarning(text,...) _Log(LOG_WARNING, text, ##__VA_ARGS__)
 
-static unsigned long __debug_tick;
+static unsigned int __debug_tick;
 
 static pthread_t PLC_thread;
 static pthread_mutex_t python_wait_mutex = PTHREAD_MUTEX_INITIALIZER;
@@ -305,9 +305,9 @@ int stopPLC()
     return 0;
 }
 
-extern unsigned long __tick;
+extern unsigned int __tick;
 
-int WaitDebugData(unsigned long *tick)
+int WaitDebugData(unsigned int *tick)
 {
     int res;
     if (PLC_shutdown) return 1;
